HEEDLS-NONE Split controller tests into multiple files (#201)
Separate out learning menu controller tests into one file per controller action being tested.
